fix: [M3-7605] - Wrong status indicator when provisioning a LKE (#10320)
## Description 📝
Fix the status icon indicator when a Kubernetes Linode is provisioning from grey to pulsing orange. I also went ahead and did some styling/file cleanup.

## How to test 🧪
### Reproduction steps
(How to reproduce the issue, if applicable)
- On develop or prod, create a new Kubernetes cluster and observe the grey provisioning Linode icon status in the details page

### Verification steps
(How to verify changes)
- Create a new Kubernetes cluster and observe the pulsing orange provisioning Linode icon status in the details page
- Ensure there are no visual regressions from the code clean up
