+++ packages/manager/.changeset/pr-10320-fixed-1711553963315.md
@@ -1,0 +1,5 @@
+---
+"@linode/manager": Fixed
+---
+
+Wrong status indicator when provisioning a LKE ([#10320](https://github.com/linode/manager/pull/10320))
